feat: restructure the directories for an app
